+++ liquid/golden/range_objects.py
@@ -21,4 +21,16 @@ cases = [
         globals={"start": 5, "end": 1},
         expect="",
     ),
+    Case(
+        description="start is negative",
+        template=r"{{ (start..end) | join: '#' }}",
+        globals={"start": -5, "end": 1},
+        expect="-5#-4#-3#-2#-1#0#1",
+    ),
+    Case(
+        description="start and end are negative",
+        template=r"{{ (start..end) | join: '#' }}",
+        globals={"start": -5, "end": -2},
+        expect="-5#-4#-3#-2",
+    ),
 ]